Refactor pomodoro and todo list
Take out models from viewmodels to their own files
