+++ src/app/viewmodel/model/Task.ts
@@ -1,0 +1,7 @@
+export default interface Task {
+  id: number;
+  title: string;
+  dueTime?: Date;
+  isDone: boolean;
+  isArchived: boolean;
+}
