+++ src/app/viewmodel/model/Time.ts
@@ -1,0 +1,5 @@
+export default interface Time {
+  h: number;
+  m: number;
+  s: number;
+};
